projects and members creating and updating

+++ database/migrations/2024_03_05_095515_create_projects_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('projects', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('owner_id');
+            $table->unsignedBigInteger('owner_id');
             $table->unsignedBigInteger('rate');
-            $table->integer('currency_id');
+            $table->unsignedBigInteger('currency_id');
             $table->timestamps();
         });
     }
